requirements.txt
django-model-utils==3.2.0
django-weasyprint==0.5.4
